+++ api/core/v2/entity_test.go
@@ -2,6 +2,7 @@ package v2
 
 import (
 	"encoding/json"
+	"reflect"
 	"sort"
 	"testing"
 
@@ -115,6 +116,35 @@ func TestSortEntitiesByLastSeen(t *testing.T) {
 		t.Run(tc.name, func(t *testing.T) {
 			sort.Sort(SortEntitiesByLastSeen(tc.inRecords))
 			assert.EqualValues(t, tc.expected, tc.inRecords)
+		})
+	}
+}
+
+func TestAddEntitySubscription(t *testing.T) {
+	tests := []struct {
+		name          string
+		entityName    string
+		subscriptions []string
+		want          []string
+	}{
+		{
+			name:          "the entity subscription is added if missing",
+			entityName:    "foo",
+			subscriptions: []string{},
+			want:          []string{"entity:foo"},
+		},
+		{
+			name:          "the entity subscription is not added if already present",
+			entityName:    "foo",
+			subscriptions: []string{"entity:foo"},
+			want:          []string{"entity:foo"},
+		},
+	}
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			if got := AddEntitySubscription(tt.entityName, tt.subscriptions); !reflect.DeepEqual(got, tt.want) {
+				t.Errorf("addEntitySubscription() = %v, want %v", got, tt.want)
+			}
 		})
 	}
 }